added heating job

+++ src/main/kotlin/com/hyblerm/homecontroller/repository/entity/OpenHabModel.kt
@@ -12,7 +12,7 @@ class OpenHabModel {
         }
 
         fun getInt(): Int {
-            return state.toInt()
+            return state.toDouble().toInt()
         }
 
         fun getDouble(): Double {
